poprawka w licezniu statystyk

+++ projekt/src/graph/Stats.kt
@@ -1,3 +1,3 @@
 package graph
 
-data class Stats(public var succeses: Int = 0, public var waits: Int = 0, public var collisions: Int = 0, public var tries: Int = 0)
+data class Stats(public var succeses: Double = 0.0, public var waits: Double = 0.0, public var collisions: Double = 0.0, public var tries: Double = 0.0)
